Corrected put_fwd install script

+++ pkt_fwd_update.sh
@@ -3,7 +3,9 @@
 wget --no-cache https://raw.githubusercontent.com/greenward/PiscesQoLDashboard/main/green_packet_v2.0.1.tgz -O /tmp/green_packet_v2.0.1.tgz
 
 tar -xzf /tmp/green_packet_v2.0.1.tgz -C /home/pi/hnt/paket
-cp /home/pi/api/tool/onPacket.sh /home/pi/api/tool/onPacket.sh.orig
+if ! test -f /home/pi/api/tool/onPacket.sh.orig; then
+  cp /home/pi/api/tool/onPacket.sh /home/pi/api/tool/onPacket.sh.orig
+fi
 sed -i 's|/home/pi/hnt/paket/paket/packet_forwarder/|/home/pi/hnt/paket/green_packet/|' /home/pi/api/tool/onPacket.sh
 kill $(ps ax | grep './lora_pkt_fwd' | grep -v 'grep' | awk '{print $1}')
